gestorh-backend/requirements.txt
Flask==3.1.2
flask-cors==5.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.0
requests==2.32.3
python-dotenv==1.0.1